Replace Validator with the new one; Remove deprecated classes from hercules-kafka-util

+++ hercules-kafka-util/src/main/java/ru/kontur/vostok/hercules/kafka/util/KafkaDefaults.java
@@ -8,5 +8,5 @@ public final class KafkaDefaults {
 
     }
 
-    public static final short DEFAULT_REPLICATION_FACTOR = 3;
+    public static final int DEFAULT_REPLICATION_FACTOR = 3;
 }
